+++ Task 4/18.fibonacci.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 int main() {
 
-    int n=0, prev = 1, ant = 0, f;
+    unsigned long long n=0, prev = 1, ant = 0, f;
 
     while(n<0){
     cout << "Enter a position of the fibonacci sequence (0 it's first position): ";
